Decouple database and repository from `shared` module
Introducing `data:local-caching` with purpose to manage
local caching data storage and retrieval. This is where
the database tables, DAOs, PO and queries can be found.

Introducing `data:repository-caching` which depends on
`remote-caching` and `local-caching` modules. It gives
access to Repository implementations which are fetching
the data from remote caching service and storing it in
the local caching service/database. This module should
be used for accessing the data from `app` and `shared`.

Refactors `shared` library to be a self-contained module
that internally manages its own dependencies and exposes
a simplified API to iOS app. This is the place where we
can put further shared code between platforms; and for
now we sharing only repositories which can be accessed
using RepositoryModule object from library.