Prepare EventRepository for decoupling from `app` module
This commit creates the necessary interfaces and domain
models for EventRepository.kt extraction from `app` to
`data:repository-caching` module and makes the changes
in EventRepository to ease the transition later.

PrimalPublisher interface has been defined in domain
which would be the dependency for all repositories
which need to publish events to nostr relays.

NostrEventPublisher, NostrEventSignatureHandler and
NostrEventImporter interface are also defined for
future use. The KMP implementation of PrimalPublisher
should depend only those 3 interfaces.