Remove PrimalDatabase from `app` module
Remove the entire PrimalDatabase from `app` and start
using only repositories from `data:repository-caching`
where we have KMP ported version of the caching db.

All files and classes which were referencing entities
from removed database were: (1) migrated to use repos
and/or domain models if the change was straightforward;
(2) commented out if the change is time consuming atm;
or (3) removed if they were obsolete or already ported
to data:repository-caching module.

NotePublishHandler has been decouple from the database
and dependency was replaced with FeedRepository and
EventRelayHintsRepository.

Breaks "New Notes Pill" feature. NoteFeedViewModel needs
missing functionality from FeedRepository. We should
reconsider the current approach for

Breaks Legend & OG leaderboards. PremiumRepository was
coupled with database and was using db entities for
processing the data. We need to write new mappers.

Breaks proper logging out. We need to implement in the
repository-caching module an api to clear the data
and use it in the UserRepository.

Breaks profile updates. Database mappers were used to
process the update and save it in the database and
AccountsStore. We need a better way of handling
profile updates.

Breaks UserAccountFetcher. It was tightly coupled with
the database and direct api access. We need to
reimplement it to use proper repositories.

Breaks caching of profiles coming from WalletTxMediator.
We no longer have direct database access. We need to
find a way to sync or reuse data between the databases.
Here are some sync/import edge cases:
- PremiumRepository uses PremiumApi gets some profiles
  which we could import in the caching database.
- UserRepository is dealing with our own profile which
  we could as well import.
- WalletTransactionsMediator receives from WalletApi
  all profiles related to transactions which needs to
  be imported in the caching database for later use.

Already broken referenced events, profiles, articles,
highlights and zaps.