Deduplicate `Tags` file
Deduplication of `Tags` required a slight refactor of `NostrResources`.
Some functions in `NostrResources` were moved to `NostrUriUtils` as they
were more fitting there and to be visible in `domain:nostr` module.
Following this change it was made obvious that we should get rid of
`NostrResources` in `app` module as it was a union of existing
`NostrResources` and `NostrUriUtils`.