Implement proof of concept Amber signing solution
Problem we were facing with Amber signing is that all signing comes down
to UI. This solution uses `MainActivity` as that UI to handle launching
sign intents and propagate results to our Notary. This should, in
theory, solve every type of sign we have indentified before embarking on
this journey.