Implement thread muting in note query builders and fix multi-account bug
Adds `isThreadMuted` field to `FeedPostData` and propagates that value
to related entities. Previous `isMuted` is now `isAuthorMuted`.

Fixes bug in which all muted items didn't recognize currently logged in
user which caused all local muting to happen on a global level. Muting
user on Account A would cause that user to be muted locally on Account B
as well.