Implement Amber signing solution (#415)
* Make Amber ContentResolver sign function return `AmberSignResult`
Previously, for the sake of simplicity, we have ignored `rejected`
column in the result. This accounts for that case.

* Implement Amber signing solution
Problem we were facing with Amber signing is that all signing comes
down to UI. This solution uses `MainActivity` as that UI to handle
launching sign intents and propagate results to our Notary. This should
solve every type of sign we have identified.

* Implement new signing logic in `NostrNotary`

* Refactor `NostrNotary` usages

* Ask for permissions to sign ApplicationSpecificData and Wallet op event

* Fix CreateAccountHandlerTests