Remove transaction when inserting conversation
Since we migrated to KMP Room messages stopped
working when fetching conversations. It seems
that database ends in a deadlock. The changes
that fixed the deadlock were to either stop
fetching referenced notes found in messages or
to stop inserting messages and conversations
in transaction.

Until we find the real cause, this commit is
going to remove the transaction inserting
of messages & conversations.