Implement topological sort for FeedPost on threads (#437)
We have relied on `createdAt` timestamp to figure out the root post and
correct chain of replies. We have seen some cases in which root post
doesn't have oldest timestamp which renders this method obsolete. This
implements topological sort that follows `reply` and `root` tags on
posts to figure out the correct order. This has the same time complexity
as the DFS and shouldn't cause performance issues.