Improve `FullScreenController`
We now remember the last `requestedOrientation` and restoring it once we
are done with `FullScreenController`. This allows other screens to keep
their preferred orientation.

`isPortraitDeg` got a sagnificant improvement once we discovered that
mapping of rotation degrees to orientation depends on the current
orientation. Android System favors the current orientation in this
resolution process.

We are now keeping track of auto rotate system setting. Small but key
change was adding `delay` before unlocking rotation to allow the device
to settle. Some weird rotation bugs occur if user rotates the device
slowly.